Remove the HTMLTable{Header,Data}CellElement interfaces
Fixes #17222.
