+++ components/script/dom/mod.rs
@@ -372,9 +372,7 @@ pub mod htmlstyleelement;
 pub mod htmltablecaptionelement;
 pub mod htmltablecellelement;
 pub mod htmltablecolelement;
-pub mod htmltabledatacellelement;
 pub mod htmltableelement;
-pub mod htmltableheadercellelement;
 pub mod htmltablerowelement;
 pub mod htmltablesectionelement;
 pub mod htmltemplateelement;
