--- components/script/dom/webidls/HTMLTableDataCellElement.webidl
@@ -1,15 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-// https://html.spec.whatwg.org/multipage/#htmltabledatacellelement
-[HTMLConstructor]
-interface HTMLTableDataCellElement : HTMLTableCellElement {
-  // also has obsolete members
-};
-
-// https://html.spec.whatwg.org/multipage/#HTMLTableDataCellElement-partial
-partial interface HTMLTableDataCellElement {
-  // [CEReactions]
-  //         attribute DOMString abbr;
-};
