--- components/script/dom/webidls/HTMLTableHeaderCellElement.webidl
@@ -1,15 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-// https://html.spec.whatwg.org/multipage/#htmltableheadercellelement
-[HTMLConstructor]
-interface HTMLTableHeaderCellElement : HTMLTableCellElement {
-  // [CEReactions]
-  //         attribute DOMString scope;
-  // [CEReactions]
-  //         attribute DOMString abbr;
-  // [CEReactions]
-  //         attribute DOMString sorted;
-  // void sort();
-};
