+++ tests/wpt/metadata/html/dom/historical.html.ini
@@ -3,9 +3,3 @@
   [document.all cannot find applet]
     expected: FAIL
 
-  [HTMLTableHeaderCellElement interface is removed]
-    expected: FAIL
-
-  [HTMLTableDataCellElement interface is removed]
-    expected: FAIL
-
