+++ tests/wpt/mozilla/meta/MANIFEST.json
@@ -26638,7 +26638,7 @@
    "testharness"
   ],
   "mozilla/collections.html": [
-   "4011ee6bf322eaacafafd98238c4261084397fde",
+   "8e06ffcc0933719b4b79ea6656d6635cc121d900",
    "testharness"
   ],
   "mozilla/createEvent-storageevent.html": [
@@ -27038,7 +27038,7 @@
    "testharness"
   ],
   "mozilla/interfaces.html": [
-   "4863f4d1036e945cfaae1f112981fac3367b2e49",
+   "8bbde3b46a8e9d5ead6f9a2bf372d9647ad059be",
    "testharness"
   ],
   "mozilla/interfaces.js": [
