+++ tests/wpt/mozilla/tests/mozilla/interfaces.html
@@ -139,9 +139,7 @@ test_interfaces([
   "HTMLTableCaptionElement",
   "HTMLTableCellElement",
   "HTMLTableColElement",
-  "HTMLTableDataCellElement",
   "HTMLTableElement",
-  "HTMLTableHeaderCellElement",
   "HTMLTableRowElement",
   "HTMLTableSectionElement",
   "HTMLTemplateElement",
